moving temporary_blob_storage into storage_client
