+++ medicines/doc-index-updater/src/pars_upload.rs
@@ -3,8 +3,7 @@ use crate::{
     document_manager::{accept_job, check_in_document_handler},
     models::{Document, FileSource},
     state_manager::{with_state, JobStatusClient, StateManager},
-    storage_client::StorageClient,
-    temporary_blob_storage::{StorageFile, TemporaryBlobStorage},
+    storage_client::{models::StorageFile, StorageClient, TemporaryBlobStorage},
 };
 use bytes::BufMut;
 use futures::future::join_all;
