+++ medicines/doc-index-updater/src/storage_client/mod.rs
@@ -1,15 +1,18 @@
-use crate::temporary_blob_storage::StorageClientError;
 use azure_sdk_core::errors::AzureError;
 use azure_sdk_storage_core::prelude::Client;
 use client::BlobClient;
+use models::StorageClientError;
 
+pub use azure_blob_client::TemporaryBlobStorage;
 pub use client::StorageClient;
 pub use delete::DeleteBlob;
 pub use get::GetBlob;
 
+mod azure_blob_client;
 mod client;
 mod delete;
 mod get;
+pub mod models;
 
 pub struct BlobResponse {
     pub blob_name: String,
