+++ medicines/doc-index-updater/src/storage_client/models.rs
@@ -1,0 +1,11 @@
+pub struct StorageFile {
+    pub name: String,
+    pub path: String,
+}
+
+#[derive(Debug)]
+pub enum StorageClientError {
+    RetrievalError { message: String },
+    UploadError { message: String },
+    ClientError { message: String },
+}
